split off system randomness functions into a new module
