+++ node-template/runtime/Cargo.toml
@@ -20,6 +20,7 @@ babe-primitives = { package = "substrate-consensus-babe-primitives", path = "../
 executive = { package = "srml-executive", path = "../../srml/executive", default_features = false }
 indices = { package = "srml-indices", path = "../../srml/indices", default_features = false }
 grandpa = { package = "srml-grandpa", path = "../../srml/grandpa", default-features = false }
+randomness = { package = "srml-randomness", path = "../../srml/randomness", default_features = false }
 system = { package = "srml-system", path = "../../srml/system", default_features = false }
 timestamp = { package = "srml-timestamp", path = "../../srml/timestamp", default_features = false }
 sudo = { package = "srml-sudo", path = "../../srml/sudo", default_features = false }
@@ -46,6 +47,7 @@ std = [
 	"grandpa/std",
 	"primitives/std",
 	"sr-primitives/std",
+	"randomness/std",
 	"system/std",
 	"timestamp/std",
 	"sudo/std",
