+++ srml/contracts/Cargo.toml
@@ -17,6 +17,7 @@ rstd = { package = "sr-std", path = "../../core/sr-std", default-features = fals
 sandbox = { package = "sr-sandbox", path = "../../core/sr-sandbox", default-features = false }
 support = { package = "srml-support", path = "../support", default-features = false }
 system = { package = "srml-system", path = "../system", default-features = false }
+randomness = { package = "srml-randomness", path = "../randomness", default-features = false }
 timestamp = { package = "srml-timestamp", path = "../timestamp", default-features = false }
 
 [dev-dependencies]
@@ -33,6 +34,7 @@ std = [
 	"codec/std",
 	"primitives/std",
 	"sr-primitives/std",
+	"randomness/std",
 	"runtime-io/std",
 	"rstd/std",
 	"sandbox/std",
