+++ srml/randomness/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "srml-randomness"
+version = "1.0.0"
+authors = ["Parity Technologies <admin@parity.io>"]
+edition = "2018"
+
+[dependencies]
+safe-mix = { version = "1.0", default-features = false}
+codec = { package = "parity-scale-codec", version = "1.0.0", default-features = false, features = ["derive"] }
+sr-primitives = { path = "../../core/sr-primitives", default-features = false }
+support = { package = "srml-support", path = "../support", default-features = false }
+system = { package = "srml-system", path = "../system", default-features = false }
+rstd = { package = "sr-std", path = "../../core/sr-std", default-features = false }
+
+[features]
+default = ["std"]
+std = [
+	"safe-mix/std",
+	"system/std",
+	"codec/std",
+	"support/std",
+	"sr-primitives/std",
+	"rstd/std",
+]
